email content
Hi Jerry, Allen,

 

I think we need to review below issue 2# further.

Below picture is a calculation of the accumulated tolerance for the gear installation. 

With the current design, in theory it is a long tolerance link between the gear center to the feed hole center on the plastic bracket (The gear sit on the nylon bushing, the bushing sit on the bearing of the motor, the bearing assembled in the motor housing, the motor housing sit on the plastic bracket), although all these components are made/assembled in tight designed tolerances, the concentricity tolerance between the gear center to the feed hole center (X+D) is more than 0.8mm (about 1.0). Can you let us know your expectation tolerance for here. If the 1.0mm is far away than your expectation, we need to improve the design for shortening/broking the current long tolerance link.

 

The solution option 1#: Making the nylon bushing hole to tight fit with the motor spindle, we push/install the bushing onto the spindle with a